requirements.txt
Flask>=2.3.0,<3.0.0
PyYAML>=6.0,<7.0
Werkzeug>=2.3.0,<3.0.0
